Add support for multiple subnets and on-demand fallback
Multiple subnets allows falling back to a different Availability Zone
when there is insufficient capacity in an AZ, subnet-id input can be
pecified as JSON Array string ie: '["subnet-1","subnet-2"]'
Subnet will be chosen at random.

After trying (and failing) all subnets with a spot request, we will now
retry all subnets again with a regular on-demand request. This means if
there is no spot capacity, we'll create an on-demand instance instead.